# docs/tonic-v4-openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "https://api.publisher.tonic.com/v4"
    }
  ],
  "info": {
    "title": "TONIC. for Publishers API",
    "description": "This is the documentation for the \"TONIC. for Publishers API\". Currently this API is in beta and is subject to change. If you have any questions or feedback, please contact your Account Manager.",
    "version": "4.16.1",
    "contact": {
      "email": "support@tonic.com"
    }
  },
...
  "paths": {
    "/account": {
      "get": {
        "operationId": "AccountGet",
        "summary": "Get current users data",
        "description": "Returns the users mail, available campaign types, first- and lastname. In case of a subuser it will also return its role and the associated rights.",
        "deprecated": false,
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "example": {
                  "data": [
                    {
                      "email": "al.dillon@example.com",
                      "firstName": "Al",
                      "lastName": "Dillon",
                      "availableCampaignTypes": [
                        "display",
                        "search",
                        "rsoc"
                      ],
                      "remainingCampaigns": {
                        "display": 22,
                        "search": 20,
                        "rsoc": 3
                      },
                      "network": "facebook",
                      "availableRsocDomains": [
                        {
                          "domain": "example.com",
                          "languages": [
                            "en",
                            "de"
                          ]
                        },
                        {
                          "domain": "example.org",
                          "languages": [
                            "fr",
                            "zh-hans"
                          ]
                        }
                      ],
                      "isSubUser": false,
                      "subUserRole": {
                        "id": 25,
                        "name": "Campaign creator",
                        "accountSettings": false,
                        "payoutSettings": false,
                        "viewRevenue": "own",
                        "createArticles": true,
                        "createCampaigns": true,
                        "viewArticles": "all",
                        "viewCampaigns": "own",
                        "apiAccess": true,
                        "created": "2022-08-28T12:29:44-07:00",
                        "subUsers": [
                          {
                            "id": 58,
                            "email": "xyz@example.com",
                            "verificationStatus": "verified",
                            "firstName": "Exepsilon",
                            "lastName": "Zet",
                            "status": "active",
                            "created": "2022-08-28T12:29:44-07:00",
                            "subUserRole": {
                              "id": 25,
                              "name": "Campaign creator",
                              "accountSettings": false,
                              "payoutSettings": false,
                              "viewRevenue": "own",
                              "createArticles": true,
                              "createCampaigns": true,
                              "viewArticles": "all",
                              "viewCampaigns": "own",
                              "apiAccess": true,
                              "created": "2022-08-28T12:29:44-07:00"
                            }
                          }
                        ]
                      },
                      "isSystemSubUser": false,
                      "parentAccountId": 123,
                      "parentAccountEmail": "al.dillon@example.com"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "operationId": "AccountPatch",
        "summary": "Update current users data",
        "description": "Updates the current users firstName and lastName.",
        "deprecated": false,
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "AccountPatchRequest",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60,
                    "description": "The new first name of the user",
                    "example": "Al"
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60,
                    "description": "The new last name of the user",
                    "example": "Dillon"
                  }
                },
                "required": [
                  "firstName",
                  "lastName"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "example": {
                  "data": [
                    {
                      "email": "al.dillon@example.com",
                      "firstName": "Al",
                      "lastName": "Dillon",
                      "availableCampaignTypes": [
                        "display",
                        "search",
                        "rsoc"
                      ],
                      "remainingCampaigns": {
                        "display": 22,
                        "search": 20,
                        "rsoc": 3
                      },
                      "network": "facebook",
                      "availableRsocDomains": [
                        {
                          "domain": "example.com",
                          "languages": [
                            "en",
                            "de"
                          ]
                        },
                        {
                          "domain": "example.org",
                          "languages": [
                            "fr",
                            "zh-hans"
                          ]
                        }
                      ],
                      "isSubUser": false,
                      "subUserRole": {
                        "id": 25,
                        "name": "Campaign creator",
                        "accountSettings": false,
                        "payoutSettings": false,
                        "viewRevenue": "own",
                        "createArticles": true,
                        "createCampaigns": true,
                        "viewArticles": "all",
                        "viewCampaigns": "own",
                        "apiAccess": true,
                        "created": "2022-08-28T12:29:44-07:00",
                        "subUsers": [
                          {
                            "id": 58,
                            "email": "xyz@example.com",
                            "verificationStatus": "verified",
                            "firstName": "Exepsilon",
                            "lastName": "Zet",
                            "status": "active",
                            "created": "2022-08-28T12:29:44-07:00",
                            "subUserRole": {
                              "id": 25,
                              "name": "Campaign creator",
                              "accountSettings": false,
                              "payoutSettings": false,
                              "viewRevenue": "own",
                              "createArticles": true,
                              "createCampaigns": true,
                              "viewArticles": "all",
                              "viewCampaigns": "own",
                              "apiAccess": true,
                              "created": "2022-08-28T12:29:44-07:00"
                            }
                          }
                        ]
                      },
                      "isSystemSubUser": false,
                      "parentAccountId": 123,
                      "parentAccountEmail": "al.dillon@example.com"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
    "/countries": {
      "get": {
        "operationId": "CountriesGet",
        "summary": "Get available Countries",
        "description": "Get a list of available countries.",
        "deprecated": false,
        "tags": [
          "Countries"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "campaignType",
            "description": "Filter countries by campaign type.<br /><i>Available Values</i> : rsoc",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "rsoc"
          },
          {
            "in": "query",
            "name": "domain",
            "description": "Filter countries by domain compatibility.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example.com"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "example": {
                  "data": [
                    {
                      "code": "US",
                      "name": "United States"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/offers": {
      "get": {
        "operationId": "OffersGet",
        "summary": "Get Offers",
        "description": "Returns a list of available offers.",
        "deprecated": false,
        "tags": [
          "Offers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "campaignType",
            "description": "Filter by campaign type.<br /><i>Available Values</i> : rsoc",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "rsoc"
          },
          {
            "in": "query",
            "name": "countryCode",
            "description": "Filter by country code.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "US"
          },
          {
            "in": "query",
            "name": "domain",
            "description": "Filter offers by domain compatibility.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example.com"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "example": {
                  "data": [
                    {
                      "id": 42,
                      "name": "Offer XYZ",
                      "vertical": {
                        "id": 76,
                        "name": "Vertical XYZ"
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiModel": {
        "title": "ApiModel",
        "type": "object"
      },
...
      "ErrorResponse": {
        "title": "ErrorResponse",
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "minimum": 400,
            "maximum": 599,
            "example": 500
          },
          "details": {
            "title": "ErrorDetails",
            "oneOf": [
              {
                "title": "ErrorDetailsSimple",
                "type": "string",
                "example": "Internal Server Error"
              },
              {
                "title": "ErrorDetailsParameter",
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "parameterName": {
                      "type": "string",
                      "example": "name"
                    },
                    "parameterLocation": {
                      "type": "string",
                      "example": "body"
                    },
                    "collectionIndex": {
                      "type": "number",
                      "description": "Index of dataset in the collection where the error occurred",
                      "example": 0,
                      "nullable": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Missing required post body parameter 'name'"
                    }
                  }
                }
              }
            ]
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "Response": {
        "title": "Response",
        "type": "object",
        "properties": {
          "data": {
            "title": "ResponseData",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApiModel"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ApiModel"
                }
              },
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "pagination": {
            "type": "object",
            "nullable": true
          },
          "sorting": {
            "type": "object",
            "nullable": true,
            "properties": {
              "orderField": {
                "type": "string",
                "example": "id"
              },
              "orderOrientation": {
                "type": "string",
                "enum": [
                  "asc",
                  "desc"
                ],
                "example": "asc"
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 400,
              "details": "Bad Request",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
...
      "404": {
        "description": "Not Found Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 404,
              "details": "Not Found Error",
              "requestId": "root=1-6335cbde-6baf0bed7a2eaecc6b1214bf"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}